chore: rename app to App
