+++ src/main.tsx
@@ -2,7 +2,7 @@ import { createTheme, MantineProvider } from '@mantine/core';
 import '@mantine/core/styles.css';
 import { StrictMode } from 'react';
 import * as ReactDOM from 'react-dom/client';
-import { App } from './app/app';
+import { App } from './app/App';
 import './styles.css';
 
 const root = ReactDOM.createRoot(
